requirements.txt
Django>=5.0
djangorestframework>=3.14
firebase-admin>=6.2
python-dotenv>=1.0
django-cors-headers>=4.3.1
requests>=2.31.0
dj-database-url>=2.3.0
gunicorn>=23.0.0
whitenoise>=6.9.0
django-anymail>=12.0
django-storages>=1.14.5
boto3>=1.36.23
botocore>=1.36.23
s3transfer>=0.11.2